requirements.txt
Django~=3.1.3
djangorestframework==3.11.*
drf_yasg~=1.17.1
django-bootstrap-form~=3.4
djangorestframework-datatables~=0.5.2
django-filebrowser~=3.13.1
python-fsutil~=0.3.0
cryptography~=3.2.1